+++ sdk/src/main/java/com/exponea/sdk/manager/FetchManager.kt
@@ -1,7 +1,5 @@
 package com.exponea.sdk.manager
 
-import com.exponea.sdk.models.Banner
-import com.exponea.sdk.models.BannerResult
 import com.exponea.sdk.models.Consent
 import com.exponea.sdk.models.CustomerIds
 import com.exponea.sdk.models.CustomerRecommendation
@@ -9,25 +7,9 @@ import com.exponea.sdk.models.CustomerRecommendationRequest
 import com.exponea.sdk.models.ExponeaProject
 import com.exponea.sdk.models.FetchError
 import com.exponea.sdk.models.InAppMessage
-import com.exponea.sdk.models.Personalization
 import com.exponea.sdk.models.Result
 
 internal interface FetchManager {
-
-    fun fetchBannerConfiguration(
-        exponeaProject: ExponeaProject,
-        customerIds: CustomerIds,
-        onSuccess: (Result<ArrayList<Personalization>>) -> Unit,
-        onFailure: (Result<FetchError>) -> Unit
-    )
-
-    fun fetchBanner(
-        exponeaProject: ExponeaProject,
-        bannerConfig: Banner,
-        onSuccess: (Result<ArrayList<BannerResult>>) -> Unit,
-        onFailure: (Result<FetchError>) -> Unit
-    )
-
     fun fetchConsents(
         exponeaProject: ExponeaProject,
         onSuccess: (Result<ArrayList<Consent>>) -> Unit,