+++ sdk/src/main/java/com/exponea/sdk/manager/FetchManagerImpl.kt
@@ -1,7 +1,5 @@
 package com.exponea.sdk.manager
 
-import com.exponea.sdk.models.Banner
-import com.exponea.sdk.models.BannerResult
 import com.exponea.sdk.models.Consent
 import com.exponea.sdk.models.CustomerIds
 import com.exponea.sdk.models.CustomerRecommendation
@@ -10,7 +8,6 @@ import com.exponea.sdk.models.CustomerRecommendationResponse
 import com.exponea.sdk.models.ExponeaProject
 import com.exponea.sdk.models.FetchError
 import com.exponea.sdk.models.InAppMessage
-import com.exponea.sdk.models.Personalization
 import com.exponea.sdk.models.Result
 import com.exponea.sdk.network.ExponeaService
 import com.exponea.sdk.util.Logger
@@ -63,36 +60,6 @@ internal class FetchManagerImpl(
                 onFailure(Result(false, error))
             }
         }
-    }
-
-    override fun fetchBannerConfiguration(
-        exponeaProject: ExponeaProject,
-        customerIds: CustomerIds,
-        onSuccess: (Result<ArrayList<Personalization>>) -> Unit,
-        onFailure: (Result<FetchError>) -> Unit
-    ) {
-        api.getBannerConfiguration(exponeaProject).enqueue(
-            getFetchCallback(
-                object : TypeToken<Result<ArrayList<Personalization>>>() {},
-                onSuccess,
-                onFailure
-            )
-        )
-    }
-
-    override fun fetchBanner(
-        exponeaProject: ExponeaProject,
-        bannerConfig: Banner,
-        onSuccess: (Result<ArrayList<BannerResult>>) -> Unit,
-        onFailure: (Result<FetchError>) -> Unit
-    ) {
-        api.postFetchBanner(exponeaProject, bannerConfig).enqueue(
-            getFetchCallback(
-                object : TypeToken<Result<ArrayList<BannerResult>>>() {},
-                onSuccess,
-                onFailure
-            )
-        )
     }
 
     override fun fetchConsents(