+++ sdk/src/main/java/com/exponea/sdk/models/ApiEndPoint.kt
@@ -9,8 +9,6 @@ internal data class ApiEndPoint(private val endPointName: EndPointName, private 
         TRACK_EVENTS,
         TRACK_CAMPAIGN,
         CUSTOMERS_ATTRIBUTES,
-        CONFIGURE_BANNER,
-        SHOW_BANNER,
         CONSENTS,
         IN_APP_MESSAGES
     }
@@ -21,8 +19,6 @@ internal data class ApiEndPoint(private val endPointName: EndPointName, private 
             EndPointName.TRACK_EVENTS -> "/track/v2/projects/$splitterToken/customers/events"
             EndPointName.TRACK_CAMPAIGN -> "/track/v2/projects/$splitterToken/campaigns/clicks"
             EndPointName.CUSTOMERS_ATTRIBUTES -> "/data/v2/projects/$splitterToken/customers/attributes"
-            EndPointName.CONFIGURE_BANNER -> "/track/v2/projects/$splitterToken/configuration/banners"
-            EndPointName.SHOW_BANNER -> "/data/v2/projects/$splitterToken/customers/personalisation/show-banners"
             EndPointName.CONSENTS -> "/data/v2/projects/$splitterToken/consent/categories"
             EndPointName.IN_APP_MESSAGES -> "/webxp/s/$splitterToken/inappmessages"
         }