+++ sdk/src/main/java/com/exponea/sdk/models/BannerResult.kt
@@ -1,7 +1,0 @@
-package com.exponea.sdk.models
-
-data class BannerResult(
-    var html: String? = null,
-    var style: String? = null,
-    var script: String? = null
-)