+++ sdk/src/main/java/com/exponea/sdk/models/TypeUrl.kt
@@ -1,6 +1,0 @@
-package com.exponea.sdk.models
-
-internal data class TypeUrl(
-    var type: String? = null,
-    var url: String? = null
-)