+++ sdk/src/test/java/com/exponea/sdk/runcatching/ExponeaExceptionThrowing.kt
@@ -21,12 +21,10 @@ object ExponeaExceptionThrowing {
         every { anyConstructed<ExponeaComponent>().sessionManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().connectionManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().flushManager } throws TestPurposeException()
-        every { anyConstructed<ExponeaComponent>().personalizationManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().backgroundTimerManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().serviceManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().fcmManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().fetchManager } throws TestPurposeException()
-        every { anyConstructed<ExponeaComponent>().fileManager } throws TestPurposeException()
         every { anyConstructed<ExponeaComponent>().networkManager } throws TestPurposeException()
 
         // This will cause onSessionStart/Stop to throw exception