+++ sdk/src/test/java/com/exponea/sdk/runcatching/PublicApiTestCases.kt
@@ -27,7 +27,8 @@ internal object PublicApiTestCases {
         Pair(Exponea::sessionTimeout, Constants.Session.defaultTimeout),
         Pair(Exponea::tokenTrackFrequency, Constants.Token.defaultTokenFrequency),
         Pair(Exponea::isInitialized, false),
-        Pair(Exponea::notificationDataCallback, null)
+        Pair(Exponea::notificationDataCallback, null),
+        Pair(Exponea::customerCookie, null)
     )
 
     val initMethods: Array<Pair<KFunction<Any>, () -> Any>> = arrayOf(
@@ -50,10 +51,6 @@ internal object PublicApiTestCases {
             { Exponea.anonymize() }
         ),
         Pair(
-            Exponea::showBanners,
-            { Exponea.showBanners(CustomerIds()) }
-        ),
-        Pair(
             Exponea::identifyCustomer,
             { Exponea.identifyCustomer(CustomerIds(), PropertiesList(hashMapOf())) }
         ),
@@ -64,10 +61,6 @@ internal object PublicApiTestCases {
         Pair(
             Exponea::getConsents,
             { Exponea.getConsents({}, {}) }
-        ),
-        Pair(
-            Exponea::getPersonalizationWebLayer,
-            { Exponea.getPersonalizationWebLayer({}, {}) }
         ),
         Pair(
             Exponea::handleCampaignIntent,