+++ sdk/src/test/java/com/exponea/sdk/telemetry/TelemetryUtilityTest.kt
@@ -194,8 +194,7 @@ internal class TelemetryUtilityTest {
                     "pushChannelName" to "Exponea [default]",
                     "automaticPushNotification" to "false",
                     "pushNotificationImportance" to "3 [default]",
-                    "sessionTimeout" to "20.0 [default]",
-                    "contentType" to "application/json [default]"
+                    "sessionTimeout" to "20.0 [default]"
                 ),
                 TelemetryUtility.formatConfigurationForTracking(ExponeaConfiguration(
                     projectToken = "mock_project_token",